db연동 완료, 차트 소수점 2자리 완료, 백업본

+++ src/main/java/com/flex/bizone/map/ChartDataResponse.java
@@ -1,68 +1,208 @@
 package com.flex.bizone.map;
 
 public class ChartDataResponse {
-    private int avgMonthlySalesScore;
-    private int totalWorkplacePopulationScore;
-    private int attractionCountScore;
-    private int totalExpenditureScore;
-    private int avgRentFeeScore;
-    private int otherScoresTotal;
-    private int successProbability;
+    private double avgMonthlySalesScore;
+    private double totalWorkplacePopulationScore;
+    private double attractionCountScore;
+    private double totalExpenditureScore;
+    private double avgRentFeeScore;
+    private double otherScoresTotal;
+    private double successProbability;
 
     // Getter와 Setter 메서드
-    public int getAvgMonthlySalesScore() {
+    public double getAvgMonthlySalesScore() {
         return avgMonthlySalesScore;
     }
 
-    public void setAvgMonthlySalesScore(int avgMonthlySalesScore) {
+    public void setAvgMonthlySalesScore(double avgMonthlySalesScore) {
         this.avgMonthlySalesScore = avgMonthlySalesScore;
     }
 
-    public int getTotalWorkplacePopulationScore() {
+    public double getTotalWorkplacePopulationScore() {
         return totalWorkplacePopulationScore;
     }
 
-    public void setTotalWorkplacePopulationScore(int totalWorkplacePopulationScore) {
+    public void setTotalWorkplacePopulationScore(double totalWorkplacePopulationScore) {
         this.totalWorkplacePopulationScore = totalWorkplacePopulationScore;
     }
 
-    public int getAttractionCountScore() {
+    public double getAttractionCountScore() {
         return attractionCountScore;
     }
 
-    public void setAttractionCountScore(int attractionCountScore) {
+    public void setAttractionCountScore(double attractionCountScore) {
         this.attractionCountScore = attractionCountScore;
     }
 
-    public int getTotalExpenditureScore() {
+    public double getTotalExpenditureScore() {
         return totalExpenditureScore;
     }
 
-    public void setTotalExpenditureScore(int totalExpenditureScore) {
+    public void setTotalExpenditureScore(double totalExpenditureScore) {
         this.totalExpenditureScore = totalExpenditureScore;
     }
 
-    public int getAvgRentFeeScore() {
+    public double getAvgRentFeeScore() {
         return avgRentFeeScore;
     }
 
-    public void setAvgRentFeeScore(int avgRentFeeScore) {
+    public void setAvgRentFeeScore(double avgRentFeeScore) {
         this.avgRentFeeScore = avgRentFeeScore;
     }
 
-    public int getOtherScoresTotal() {
+    public double getOtherScoresTotal() {
         return otherScoresTotal;
     }
 
-    public void setOtherScoresTotal(int otherScoresTotal) {
+    public void setOtherScoresTotal(double otherScoresTotal) {
         this.otherScoresTotal = otherScoresTotal;
     }
 
-    public int getSuccessProbability() {
+    public double getSuccessProbability() {
         return successProbability;
     }
 
-    public void setSuccessProbability(int successProbability) {
+    public void setSuccessProbability(double successProbability) {
         this.successProbability = successProbability;
     }
 }
+
+// 실험2
+//package com.flex.bizone.map;
+//
+//public class ChartDataResponse {
+//    private float avgMonthlySalesScore;
+//    private float totalWorkplacePopulationScore;
+//    private float attractionCountScore;
+//    private float totalExpenditureScore;
+//    private float avgRentFeeScore;
+//    private float otherScoresTotal;
+//    private float successProbability;
+//
+//    // Getter와 Setter 메서드
+//    public float getAvgMonthlySalesScore() {
+//        return avgMonthlySalesScore;
+//    }
+//
+//    public void setAvgMonthlySalesScore(float avgMonthlySalesScore) {
+//        this.avgMonthlySalesScore = avgMonthlySalesScore;
+//    }
+//
+//    public float getTotalWorkplacePopulationScore() {
+//        return totalWorkplacePopulationScore;
+//    }
+//
+//    public void setTotalWorkplacePopulationScore(float totalWorkplacePopulationScore) {
+//        this.totalWorkplacePopulationScore = totalWorkplacePopulationScore;
+//    }
+//
+//    public float getAttractionCountScore() {
+//        return attractionCountScore;
+//    }
+//
+//    public void setAttractionCountScore(float attractionCountScore) {
+//        this.attractionCountScore = attractionCountScore;
+//    }
+//
+//    public float getTotalExpenditureScore() {
+//        return totalExpenditureScore;
+//    }
+//
+//    public void setTotalExpenditureScore(float totalExpenditureScore) {
+//        this.totalExpenditureScore = totalExpenditureScore;
+//    }
+//
+//    public float getAvgRentFeeScore() {
+//        return avgRentFeeScore;
+//    }
+//
+//    public void setAvgRentFeeScore(float avgRentFeeScore) {
+//        this.avgRentFeeScore = avgRentFeeScore;
+//    }
+//
+//    public float getOtherScoresTotal() {
+//        return otherScoresTotal;
+//    }
+//
+//    public void setOtherScoresTotal(float otherScoresTotal) {
+//        this.otherScoresTotal = otherScoresTotal;
+//    }
+//
+//    public float getSuccessProbability() {
+//        return successProbability;
+//    }
+//
+//    public void setSuccessProbability(float successProbability) {
+//        this.successProbability = successProbability;
+//    }
+//}
+
+
+//package com.flex.bizone.map;
+//
+//public class ChartDataResponse {
+//    private int avgMonthlySalesScore;
+//    private int totalWorkplacePopulationScore;
+//    private int attractionCountScore;
+//    private int totalExpenditureScore;
+//    private int avgRentFeeScore;
+//    private int otherScoresTotal;
+//    private int successProbability;
+//
+//    // Getter와 Setter 메서드
+//    public int getAvgMonthlySalesScore() {
+//        return avgMonthlySalesScore;
+//    }
+//
+//    public void setAvgMonthlySalesScore(int avgMonthlySalesScore) {
+//        this.avgMonthlySalesScore = avgMonthlySalesScore;
+//    }
+//
+//    public int getTotalWorkplacePopulationScore() {
+//        return totalWorkplacePopulationScore;
+//    }
+//
+//    public void setTotalWorkplacePopulationScore(int totalWorkplacePopulationScore) {
+//        this.totalWorkplacePopulationScore = totalWorkplacePopulationScore;
+//    }
+//
+//    public int getAttractionCountScore() {
+//        return attractionCountScore;
+//    }
+//
+//    public void setAttractionCountScore(int attractionCountScore) {
+//        this.attractionCountScore = attractionCountScore;
+//    }
+//
+//    public int getTotalExpenditureScore() {
+//        return totalExpenditureScore;
+//    }
+//
+//    public void setTotalExpenditureScore(int totalExpenditureScore) {
+//        this.totalExpenditureScore = totalExpenditureScore;
+//    }
+//
+//    public int getAvgRentFeeScore() {
+//        return avgRentFeeScore;
+//    }
+//
+//    public void setAvgRentFeeScore(int avgRentFeeScore) {
+//        this.avgRentFeeScore = avgRentFeeScore;
+//    }
+//
+//    public int getOtherScoresTotal() {
+//        return otherScoresTotal;
+//    }
+//
+//    public void setOtherScoresTotal(int otherScoresTotal) {
+//        this.otherScoresTotal = otherScoresTotal;
+//    }
+//
+//    public int getSuccessProbability() {
+//        return successProbability;
+//    }
+//
+//    public void setSuccessProbability(int successProbability) {
+//        this.successProbability = successProbability;
+//    }
+//}
